Unit testing update

+++ R/internals.R
@@ -8,10 +8,13 @@
 #' @return If the file exists and is not stale, returns TRUE, otherwise FALSE.
 #'
 isValidCachedObject <- function(x){
+  
+  if ( is.null(getOption("maxCacheAge")) ) { return(FALSE) } else {
   is_there<-file.exists(x)
   #Age of files in days
   is_fresh<-as.numeric(Sys.time()-file.info(x)$mtime ) < getOption("maxCacheAge")
-  return (is_there && is_fresh)
+  return (is_there && is_fresh) }
+  
 }
 
 #' @export
@@ -24,11 +27,13 @@ isValidCachedObject <- function(x){
 #' @return Returns a cached data object.
 #'
 getCachedObject<-function(sig,wd=paste0(getwd(),"/.R_Cache/")) {
+  if ( is.null(getOption("maxCacheAge")) ) { return(NULL) } else {
   this_obj = file=paste0(getwd(),"/.R_Cache/",sig)
   if (isValidCachedObject(this_obj)) { return(readRDS(this_obj)) }
   else {
     unlink(this_obj)
     return(NULL)}
+  }
 }
 
 
@@ -45,10 +50,14 @@ getCachedObject<-function(sig,wd=paste0(getwd(),"/.R_Cache/")) {
 #'
 
 
-saveCachedObject<-function(toCache,sig,wd=paste0(getwd(),"/.R_Cache/")) {
-  if (!file.exists(wd)) {dir.create(wd)}
-  saveRDS(toCache,file=paste0(wd,sig))
-}
+saveCachedObject <-
+  function(toCache, sig, wd = paste0(getwd(), "/.R_Cache/")) {
+    if (!is.null(getOption("maxCacheAge"))) {
+    if (!file.exists(wd)) {
+      dir.create(wd)
+    }
+    saveRDS(toCache, file = paste0(wd, sig)) }
+  }
 
 #' @export
 #' @title clearCache(wd)
@@ -59,16 +68,40 @@ saveCachedObject<-function(toCache,sig,wd=paste0(getwd(),"/.R_Cache/")) {
 #' @param force Boolean (default FALSE) to clear cache without asking the user
 #' @return Nothing.
 #'
-clearCache<-function(cache_dir=paste0(getwd(),"/.R_Cache/"),force=NA) {
-  
-  
-  if ( is.na(force) ) {
-resp<-readline(paste("Are you sure you want to delete the directory ", cache_dir,"[Y/N]")) } else {
-  resp<-ifelse(force,"Y","N")
-}
-if (resp == "Y" ) { unlink(cache_dir,recursive=TRUE) } else { print("Aborting.") }
-}
+clearCache <-
+  function(cache_dir = paste0(getwd(), "/.R_Cache/"),
+           force = NA) {
+    if (is.na(force)) {
+      resp <-
+        readline(paste(
+          "Are you sure you want to delete the directory ",
+          cache_dir,
+          "[Y/N]"
+        ))
+    } else {
+      resp <- ifelse(force, "Y", "N")
+    }
+    if (resp == "Y") {
+      unlink(cache_dir, recursive = TRUE)
+    } else {
+      print("Aborting.")
+    }
+  }
 
+#' Title
+#'
+#' @return Version of the API
+#' @export
+#'
+#' 
+api_version<-function() { "27" }
+
+#' Title 
+#'
+#' @return Returns a password
+#' @export
+#'
+#' 
 get_password <- function() {
   cat("Password: ")
   system("stty -echo")
